pyproject.toml
[project]
name = "agenticragsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.77.0",
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "wikipedia>=1.4.0",
]
